+++ CHANGELOG.md
@@ -1,11 +1,22 @@
 # Release Notes
 
-## Version v0.0.8 - 2020-07-00
+## Version v0.0.8 - 2020-07-17
+
+Features:
+
+- Added kong plugin declaration on stack.yml
+- Enabled creation of files from different environments
+
+Improvements:
+
+- Removed the kong command
+
+## Version v0.0.8 - 2020-07-10
 
 Features:
 
 - Added configuration of environment variables in stack.yml
-- Added new command `sapphire function remove`
+- Added new command `safira function remove`
 
 Improvements: